+++ src/controllers/controller.ts
@@ -1,18 +1,20 @@
 import axios from 'axios';
 import { Router, type Request, type Response } from 'express';
 import { DateTime } from 'luxon';
-import { convertToReturnDetails, vatusaApi } from '../app.js';
-import { sendMail } from '../mailer.js';
+import { convertToReturnDetails } from '../app.js';
+import { sendMail } from '../helpers/mailer.js';
+import { uploadToS3 } from '../helpers/s3.js';
+import { vatusaApi } from '../helpers/vatusa.js';
 import { hasRole, isManagement, isStaff } from '../middleware/auth.js';
 import internalAuth from '../middleware/internalAuth.js';
 import getUser from '../middleware/user.js';
 import { AbsenceModel } from '../models/absence.js';
 import { ControllerHoursModel } from '../models/controllerHours.js';
+import { DossierModel } from '../models/dossier.js';
 import { NotificationModel } from '../models/notification.js';
 import { RoleModel } from '../models/role.js';
 import { UserModel } from '../models/user.js';
 import { VisitApplicationModel } from '../models/visitApplication.js';
-import { uploadToS3 } from '../s3.js';
 
 const router = Router();
 
@@ -277,7 +279,7 @@ router.post('/absence', getUser, isManagement, async (req: Request, res: Respons
 			})}</b>.`,
 		});
 
-		await req.app.dossier.create({
+		await DossierModel.create({
 			by: req.user!.cid,
 			affected: req.body.controller,
 			action: `%b added a leave of absence for %a: ${req.body.reason}`,
@@ -309,7 +311,7 @@ router.delete('/absence/:id', getUser, isManagement, async (req: Request, res: R
 
 		await absence.delete();
 
-		await req.app.dossier.create({
+		await DossierModel.create({
 			by: req.user!.cid,
 			affected: absence.controller,
 			action: `%b deleted the leave of absence for %a.`,
@@ -325,11 +327,10 @@ router.delete('/absence/:id', getUser, isManagement, async (req: Request, res: R
 router.get('/log', getUser, isStaff, async (req: Request, res: Response) => {
 	const page = +(req.query['page'] as string) || 1;
 	const limit = +(req.query['limit'] as string) || 20;
-	const amount = await req.app.dossier.countDocuments();
+	const amount = await DossierModel.countDocuments();
 
 	try {
-		const dossier = await req.app.dossier
-			.find()
+		const dossier = await DossierModel.find()
 			.sort({
 				createdAt: 'desc',
 			})
@@ -428,7 +429,7 @@ router.put('/:cid/rating', internalAuth, async (req: Request, res: Response) => 
 
 			await user.save();
 
-			await req.app.dossier.create({
+			await DossierModel.create({
 				by: -1,
 				affected: req.params['cid'],
 				action: `%a was set as Rating ${req.body.rating} by an external service.`,
@@ -617,7 +618,7 @@ router.put('/visit/:cid', getUser, hasRole(['atm', 'datm']), async (req, res) =>
 			},
 		});
 
-		await req.app.dossier.create({
+		await DossierModel.create({
 			by: req.user!.cid,
 			affected: user.cid,
 			action: `%b approved the visiting application for %a.`,
@@ -664,7 +665,7 @@ router.delete(
 				},
 			});
 
-			await req.app.dossier.create({
+			await DossierModel.create({
 				by: req.user!.cid,
 				affected: user.cid,
 				action: `%b rejected the visiting application for %a: ${req.body.reason}`,
@@ -748,7 +749,7 @@ router.post('/:cid', internalAuth, async (req: Request, res: Response) => {
 			},
 		});
 
-		await req.app.dossier.create({
+		await DossierModel.create({
 			by: -1,
 			affected: req.body.cid,
 			action: `%a was created by an external service.`,
@@ -828,7 +829,7 @@ router.put('/:cid/member', internalAuth, async (req: Request, res: Response) => 
 			});
 		}
 
-		await req.app.dossier.create({
+		await DossierModel.create({
 			by: -1,
 			affected: req.params['cid'],
 			action: `%a was ${req.body.member ? 'added to' : 'removed from'} the roster by an external service.`,
@@ -856,7 +857,7 @@ router.put('/:cid/visit', internalAuth, async (req: Request, res: Response) => {
 
 		await user.save();
 
-		await req.app.dossier.create({
+		await DossierModel.create({
 			by: -1,
 			affected: req.params['cid'],
 			action: `%a was set as a ${req.body.vis ? 'visiting controller' : 'home controller'} by an external service.`,
@@ -952,7 +953,7 @@ router.put(
 			).exec();
 
 			// Log the update in the user's dossier
-			await req.app.dossier.create({
+			await DossierModel.create({
 				by: req.user!.cid,
 				affected: req.params['cid'],
 				action: `%a was updated by %b.`,
@@ -1034,7 +1035,7 @@ router.delete('/:cid', getUser, hasRole(['atm', 'datm']), async (req: Request, r
 			});
 		}
 
-		await req.app.dossier.create({
+		await DossierModel.create({
 			by: req.user!.cid,
 			affected: req.params['cid'],
 			action: `%a was removed from the roster by %b: ${req.body.reason}`,
