+++ src/controllers/ids.ts
@@ -2,10 +2,10 @@ import axios from 'axios';
 import { Router, type Request, type Response } from 'express';
 import { Redis } from 'ioredis';
 import { convertToReturnDetails } from '../app.js';
+import zau from '../helpers/zau.js';
 import { ConfigModel } from '../models/config.js';
 import { PirepModel } from '../models/pirep.js';
 import { UserModel } from '../models/user.js';
-import zau from '../zau.js';
 
 const router = Router();
